+++ src/Viewer/CHANGELOG.md
@@ -1,4 +1,7 @@
 # 1.6.5
+* Added optional "direction cones" to cartoons visualization.
+
+# 1.6.5
 * Added BGSU RNA Loop Annotation.
 
 # 1.6.4